- restructure logic into files - implement PR fetching logic
