+++ action.yml
@@ -9,4 +9,4 @@ inputs:
     description: 'The repository token GITHUB_TOKEN'
 runs:
   using: 'node12'
-  main: 'index.js'
+  main: 'src/index.js'
