+++ package.json
@@ -2,7 +2,7 @@
   "name": "spike-retro-actions",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
